Breaks shapes into a more organized folder structure
