--- js/shapes.js
@@ -1,188 +1,0 @@
-//** Shapes **//
-
-/* UserShip */
-/*
-        •
-*/
-export const userShip = [
-  { row: 0, column: 2, color: "bg-yellow" },
-  { row: 1, column: 1, color: "bg-yellow" },
-  { row: 1, column: 2, color: "bg-yellow" },
-  { row: 1, column: 3, color: "bg-yellow" },
-  { row: 2, column: 1, color: "bg-yellow" },
-  { row: 2, column: 2, color: "bg-yellow" },
-  { row: 2, column: 3, color: "bg-yellow" },
-  { row: 3, column: 0, color: "bg-yellow" },
-  { row: 3, column: 2, color: "bg-yellow" },
-  { row: 3, column: 4, color: "bg-yellow" },
-]
-
-/* Enemies */
-/*
-        •••
-         •••
-        •••
-*/
-export const shapeSmallEnemy = [
-  {row: 0, column: 0, color: "bg-orange"},
-  {row: 0, column: 1, color: "bg-orange"},
-  {row: 0, column: 2, color: "bg-orange"},
-  {row: 1, column: 1, color: "bg-orange"},
-  {row: 1, column: 2, color: "bg-orange"},
-  {row: 1, column: 3, color: "bg-orange"},
-  {row: 2, column: 0, color: "bg-orange"},
-  {row: 2, column: 1, color: "bg-orange"},
-  {row: 2, column: 2, color: "bg-orange"}
-]
-
-/*
-         •••
-        ••
-        ••
-         •••
-*/
-export const shapeLargeEnemy = [
-  {row: 0, column: 1, color: "bg-green"},
-  {row: 0, column: 2, color: "bg-green"},
-  {row: 0, column: 3, color: "bg-green"},
-  {row: 1, column: 0, color: "bg-green"},
-  {row: 1, column: 1, color: "bg-green"},
-  {row: 2, column: 0, color: "bg-green"},
-  {row: 2, column: 1, color: "bg-green"},
-  {row: 3, column: 1, color: "bg-green"},
-  {row: 3, column: 2, color: "bg-green"},
-  {row: 3, column: 3, color: "bg-green"}
-]
-
-export const gameOver = [
-  {row: 0, column: 0, color: "bg-green"},
-  {row: 0, column: 1, color: "bg-green"},
-  {row: 0, column: 2, color: "bg-green"},
-  {row: 0, column: 3, color: "bg-green"},
-  {row: 0, column: 6, color: "bg-green"},
-  {row: 0, column: 7, color: "bg-green"},
-  {row: 0, column: 8, color: "bg-green"},
-  {row: 0, column: 9, color: "bg-green"},
-  {row: 1, column: 0, color: "bg-green"},
-  {row: 1, column: 6, color: "bg-green"},
-  {row: 1, column: 9, color: "bg-green"},
-  {row: 2, column: 0, color: "bg-green"},
-  {row: 2, column: 6, color: "bg-green"},
-  {row: 2, column: 9, color: "bg-green"},
-  {row: 3, column: 0, color: "bg-green"},
-  {row: 3, column: 2, color: "bg-green"},
-  {row: 3, column: 3, color: "bg-green"},
-  {row: 3, column: 6, color: "bg-green"},
-  {row: 3, column: 9, color: "bg-green"},
-  {row: 4, column: 0, color: "bg-green"},
-  {row: 4, column: 3, color: "bg-green"},
-  {row: 4, column: 6, color: "bg-green"},
-  {row: 4, column: 9, color: "bg-green"},
-  {row: 5, column: 0, color: "bg-green"},
-  {row: 5, column: 1, color: "bg-green"},
-  {row: 5, column: 2, color: "bg-green"},
-  {row: 5, column: 3, color: "bg-green"},
-  {row: 5, column: 6, color: "bg-green"},
-  {row: 5, column: 7, color: "bg-green"},
-  {row: 5, column: 8, color: "bg-green"},
-  {row: 5, column: 9, color: "bg-green"}
-]
-
-export const crown = [
-  {row: 0, column: 6, color: "bg-yellow"},
-  {row: 0, column: 7, color: "bg-yellow"},
-  {row: 1, column: 1, color: "bg-yellow"},
-  {row: 1, column: 5, color: "bg-yellow"},
-  {row: 1, column: 6, color: "bg-yellow"},
-  {row: 1, column: 7, color: "bg-yellow"},
-  {row: 1, column: 8, color: "bg-yellow"},
-  {row: 1, column: 12, color: "bg-yellow"},
-  {row: 2, column: 0, color: "bg-yellow"},
-  {row: 2, column: 1, color: "bg-yellow"},
-  {row: 2, column: 2, color: "bg-yellow"},
-  {row: 2, column: 6, color: "bg-yellow"},
-  {row: 2, column: 7, color: "bg-yellow"},
-  {row: 2, column: 11, color: "bg-yellow"},
-  {row: 2, column: 12, color: "bg-yellow"},
-  {row: 2, column: 13, color: "bg-yellow"},
-  {row: 3, column: 1, color: "bg-yellow"},
-  {row: 3, column: 5, color: "bg-yellow"},
-  {row: 3, column: 6, color: "bg-yellow"},
-  {row: 3, column: 7, color: "bg-yellow"},
-  {row: 3, column: 8, color: "bg-yellow"},
-  {row: 3, column: 12, color: "bg-yellow"},
-  {row: 4, column: 2, color: "bg-yellow"},
-  {row: 4, column: 3, color: "bg-yellow"},
-  {row: 4, column: 4, color: "bg-yellow"},
-  {row: 4, column: 5, color: "bg-yellow"},
-  {row: 4, column: 6, color: "bg-yellow"},
-  {row: 4, column: 7, color: "bg-yellow"},
-  {row: 4, column: 8, color: "bg-yellow"},
-  {row: 4, column: 9, color: "bg-yellow"},
-  {row: 4, column: 10, color: "bg-yellow"},
-  {row: 4, column: 11, color: "bg-yellow"},
-  {row: 5, column: 3, color: "bg-yellow"},
-  {row: 5, column: 4, color: "bg-yellow"},
-  {row: 5, column: 5, color: "bg-yellow"},
-  {row: 5, column: 6, color: "bg-yellow"},
-  {row: 5, column: 7, color: "bg-yellow"},
-  {row: 5, column: 8, color: "bg-yellow"},
-  {row: 5, column: 9, color: "bg-yellow"},
-  {row: 5, column: 10, color: "bg-yellow"},
-  {row: 6, column: 3, color: "bg-yellow"},
-  {row: 6, column: 4, color: "bg-yellow"},
-  {row: 6, column: 5, color: "bg-yellow"},
-  {row: 6, column: 6, color: "bg-yellow"},
-  {row: 6, column: 7, color: "bg-yellow"},
-  {row: 6, column: 8, color: "bg-yellow"},
-  {row: 6, column: 9, color: "bg-yellow"},
-  {row: 6, column: 10, color: "bg-yellow"},
-  {row: 8, column: 3, color: "bg-yellow"},
-  {row: 8, column: 4, color: "bg-yellow"},
-  {row: 8, column: 5, color: "bg-yellow"},
-  {row: 8, column: 6, color: "bg-yellow"},
-  {row: 8, column: 7, color: "bg-yellow"},
-  {row: 8, column: 8, color: "bg-yellow"},
-  {row: 8, column: 9, color: "bg-yellow"},
-  {row: 8, column: 10, color: "bg-yellow"},
-]
-
-/* Patterns */
-export const patternStatic = [
-  {row: 0, column: 0}
-]
-export const patternDropDown = [
-  {row: 10, column: 0}
-]
-export const pattern1 = [
-  {row: 0, column: -10},
-  {row: -5, column: 0},
-  {row: 0, column: 10},
-  {row: 5, column: 0}
-]
-export const pattern2 = [
-  {row: 4, column: 3},
-  {row: -4, column: 0},
-  {row: 6, column: 2},
-  {row: -3, column: 6}
-]
-
-export const patternScroll = [
-  { row: 1, column: 0 },
-]
-
-/* Alphabet */
-export const letterA = [
-  { row: 0, colum: 1, color: "bg-purple" },
-  { row: 0, colum: 2, color: "bg-purple" },
-  { row: 1, colum: 0, color: "bg-purple" },
-  { row: 1, colum: 3, color: "bg-purple" },
-  { row: 2, colum: 0, color: "bg-purple" },
-  { row: 2, colum: 1, color: "bg-purple" },
-  { row: 2, colum: 2, color: "bg-purple" },
-  { row: 2, colum: 3, color: "bg-purple" },
-  { row: 3, colum: 0, color: "bg-purple" },
-  { row: 3, colum: 3, color: "bg-purple" },
-  { row: 4, colum: 0, color: "bg-purple" },
-  { row: 4, colum: 3, color: "bg-purple" },
-]
