+++ js/shapes/alphabet.js
@@ -1,0 +1,15 @@
+/* Alphabet */
+export const letterA = [
+  { row: 0, colum: 1, color: "bg-purple" },
+  { row: 0, colum: 2, color: "bg-purple" },
+  { row: 1, colum: 0, color: "bg-purple" },
+  { row: 1, colum: 3, color: "bg-purple" },
+  { row: 2, colum: 0, color: "bg-purple" },
+  { row: 2, colum: 1, color: "bg-purple" },
+  { row: 2, colum: 2, color: "bg-purple" },
+  { row: 2, colum: 3, color: "bg-purple" },
+  { row: 3, colum: 0, color: "bg-purple" },
+  { row: 3, colum: 3, color: "bg-purple" },
+  { row: 4, colum: 0, color: "bg-purple" },
+  { row: 4, colum: 3, color: "bg-purple" },
+]
