+++ js/shapes/index.js
@@ -1,0 +1,49 @@
+//** Shapes **//
+
+export {
+  letterA,
+  // letterB,
+  // letterC,
+  // letterD,
+  // letterE,
+  // letterF,
+  // letterG,
+  // letterH,
+  // letterI,
+  // letterJ,
+  // letterK,
+  // letterL,
+  // letterM,
+  // letterN,
+  // letterO,
+  // letterP,
+  // letterQ,
+  // letterR,
+  // letterS,
+  // letterT,
+  // letterU,
+  // letterV,
+  // letterW,
+  // letterX,
+  // letterY,
+  // letterZ,
+} from './alphabet.js';
+
+export {
+  largeEnemyShip,
+  smallEnemyShip,
+  userShip,
+} from './ships.js';
+
+export {
+  crown,
+  gameOVer,
+} from './props.js';
+
+export {
+  pattern1,
+  pattern2,
+  patternDropdown,
+  patternScroll,
+  patternStatic,
+} from './patterns.js';
