+++ js/shapes/patterns.js
@@ -1,0 +1,25 @@
+export const pattern1 = [
+  {row: 0, column: -10},
+  {row: -5, column: 0},
+  {row: 0, column: 10},
+  {row: 5, column: 0}
+]
+
+export const pattern2 = [
+  {row: 4, column: 3},
+  {row: -4, column: 0},
+  {row: 6, column: 2},
+  {row: -3, column: 6}
+]
+
+export const patternDropDown = [
+  {row: 10, column: 0}
+]
+
+export const patternScroll = [
+  { row: 1, column: 0 },
+]
+
+export const patternStatic = [
+  {row: 0, column: 0}
+]
