+++ js/shapes/props.js
@@ -1,0 +1,92 @@
+export const crown = [
+  {row: 0, column: 6, color: "bg-yellow"},
+  {row: 0, column: 7, color: "bg-yellow"},
+  {row: 1, column: 1, color: "bg-yellow"},
+  {row: 1, column: 5, color: "bg-yellow"},
+  {row: 1, column: 6, color: "bg-yellow"},
+  {row: 1, column: 7, color: "bg-yellow"},
+  {row: 1, column: 8, color: "bg-yellow"},
+  {row: 1, column: 12, color: "bg-yellow"},
+  {row: 2, column: 0, color: "bg-yellow"},
+  {row: 2, column: 1, color: "bg-yellow"},
+  {row: 2, column: 2, color: "bg-yellow"},
+  {row: 2, column: 6, color: "bg-yellow"},
+  {row: 2, column: 7, color: "bg-yellow"},
+  {row: 2, column: 11, color: "bg-yellow"},
+  {row: 2, column: 12, color: "bg-yellow"},
+  {row: 2, column: 13, color: "bg-yellow"},
+  {row: 3, column: 1, color: "bg-yellow"},
+  {row: 3, column: 5, color: "bg-yellow"},
+  {row: 3, column: 6, color: "bg-yellow"},
+  {row: 3, column: 7, color: "bg-yellow"},
+  {row: 3, column: 8, color: "bg-yellow"},
+  {row: 3, column: 12, color: "bg-yellow"},
+  {row: 4, column: 2, color: "bg-yellow"},
+  {row: 4, column: 3, color: "bg-yellow"},
+  {row: 4, column: 4, color: "bg-yellow"},
+  {row: 4, column: 5, color: "bg-yellow"},
+  {row: 4, column: 6, color: "bg-yellow"},
+  {row: 4, column: 7, color: "bg-yellow"},
+  {row: 4, column: 8, color: "bg-yellow"},
+  {row: 4, column: 9, color: "bg-yellow"},
+  {row: 4, column: 10, color: "bg-yellow"},
+  {row: 4, column: 11, color: "bg-yellow"},
+  {row: 5, column: 3, color: "bg-yellow"},
+  {row: 5, column: 4, color: "bg-yellow"},
+  {row: 5, column: 5, color: "bg-yellow"},
+  {row: 5, column: 6, color: "bg-yellow"},
+  {row: 5, column: 7, color: "bg-yellow"},
+  {row: 5, column: 8, color: "bg-yellow"},
+  {row: 5, column: 9, color: "bg-yellow"},
+  {row: 5, column: 10, color: "bg-yellow"},
+  {row: 6, column: 3, color: "bg-yellow"},
+  {row: 6, column: 4, color: "bg-yellow"},
+  {row: 6, column: 5, color: "bg-yellow"},
+  {row: 6, column: 6, color: "bg-yellow"},
+  {row: 6, column: 7, color: "bg-yellow"},
+  {row: 6, column: 8, color: "bg-yellow"},
+  {row: 6, column: 9, color: "bg-yellow"},
+  {row: 6, column: 10, color: "bg-yellow"},
+  {row: 8, column: 3, color: "bg-yellow"},
+  {row: 8, column: 4, color: "bg-yellow"},
+  {row: 8, column: 5, color: "bg-yellow"},
+  {row: 8, column: 6, color: "bg-yellow"},
+  {row: 8, column: 7, color: "bg-yellow"},
+  {row: 8, column: 8, color: "bg-yellow"},
+  {row: 8, column: 9, color: "bg-yellow"},
+  {row: 8, column: 10, color: "bg-yellow"},
+]
+
+export const gameOver = [
+  {row: 0, column: 0, color: "bg-green"},
+  {row: 0, column: 1, color: "bg-green"},
+  {row: 0, column: 2, color: "bg-green"},
+  {row: 0, column: 3, color: "bg-green"},
+  {row: 0, column: 6, color: "bg-green"},
+  {row: 0, column: 7, color: "bg-green"},
+  {row: 0, column: 8, color: "bg-green"},
+  {row: 0, column: 9, color: "bg-green"},
+  {row: 1, column: 0, color: "bg-green"},
+  {row: 1, column: 6, color: "bg-green"},
+  {row: 1, column: 9, color: "bg-green"},
+  {row: 2, column: 0, color: "bg-green"},
+  {row: 2, column: 6, color: "bg-green"},
+  {row: 2, column: 9, color: "bg-green"},
+  {row: 3, column: 0, color: "bg-green"},
+  {row: 3, column: 2, color: "bg-green"},
+  {row: 3, column: 3, color: "bg-green"},
+  {row: 3, column: 6, color: "bg-green"},
+  {row: 3, column: 9, color: "bg-green"},
+  {row: 4, column: 0, color: "bg-green"},
+  {row: 4, column: 3, color: "bg-green"},
+  {row: 4, column: 6, color: "bg-green"},
+  {row: 4, column: 9, color: "bg-green"},
+  {row: 5, column: 0, color: "bg-green"},
+  {row: 5, column: 1, color: "bg-green"},
+  {row: 5, column: 2, color: "bg-green"},
+  {row: 5, column: 3, color: "bg-green"},
+  {row: 5, column: 6, color: "bg-green"},
+  {row: 5, column: 7, color: "bg-green"},
+  {row: 5, column: 8, color: "bg-green"},
+  {row: 5, column: 9, color: "bg-green"}
+]
