+++ js/shapes/ships.js
@@ -1,0 +1,37 @@
+export const largeEnemyShip = [
+  {row: 0, column: 1, color: "bg-green"},
+  {row: 0, column: 2, color: "bg-green"},
+  {row: 0, column: 3, color: "bg-green"},
+  {row: 1, column: 0, color: "bg-green"},
+  {row: 1, column: 1, color: "bg-green"},
+  {row: 2, column: 0, color: "bg-green"},
+  {row: 2, column: 1, color: "bg-green"},
+  {row: 3, column: 1, color: "bg-green"},
+  {row: 3, column: 2, color: "bg-green"},
+  {row: 3, column: 3, color: "bg-green"}
+]
+
+export const smallEnemyShip = [
+  {row: 0, column: 0, color: "bg-orange"},
+  {row: 0, column: 1, color: "bg-orange"},
+  {row: 0, column: 2, color: "bg-orange"},
+  {row: 1, column: 1, color: "bg-orange"},
+  {row: 1, column: 2, color: "bg-orange"},
+  {row: 1, column: 3, color: "bg-orange"},
+  {row: 2, column: 0, color: "bg-orange"},
+  {row: 2, column: 1, color: "bg-orange"},
+  {row: 2, column: 2, color: "bg-orange"}
+]
+
+export const userShip = [
+  { row: 0, column: 2, color: "bg-yellow" },
+  { row: 1, column: 1, color: "bg-yellow" },
+  { row: 1, column: 2, color: "bg-yellow" },
+  { row: 1, column: 3, color: "bg-yellow" },
+  { row: 2, column: 1, color: "bg-yellow" },
+  { row: 2, column: 2, color: "bg-yellow" },
+  { row: 2, column: 3, color: "bg-yellow" },
+  { row: 3, column: 0, color: "bg-yellow" },
+  { row: 3, column: 2, color: "bg-yellow" },
+  { row: 3, column: 4, color: "bg-yellow" },
+]
